Consolidate entry list components into shared abstractions
Extract common code from EntryList/SavedArticleList into shared components:
- ArticleListItem: unified list item with normalized props
- ArticleListStates: shared empty, error, loading, and end states
- formatRelativeTime/getDomain utilities in lib/format.ts

EntryListItem and SavedArticleListItem are now thin wrappers that
adapt their data shapes to the shared ArticleListItem component.
