+++ src/components/saved/SavedArticleList.tsx
@@ -11,6 +11,13 @@ import { useEffect, useRef, useCallback, useMemo } from "react";
 import { trpc } from "@/lib/trpc/client";
 import { SavedArticleListItem, type SavedArticleListItemData } from "./SavedArticleListItem";
 import { EntryListSkeleton } from "@/components/entries/EntryListSkeleton";
+import {
+  ArticleListEmpty,
+  ArticleListError,
+  ArticleListLoadingMore,
+  ArticleListEnd,
+  BookmarkEmptyIcon,
+} from "@/components/articles/ArticleListStates";
 
 /**
  * Filter options for the saved article list.
@@ -69,98 +76,6 @@ interface SavedArticleListProps {
    * Used by parent components for keyboard navigation and actions.
    */
   onArticlesLoaded?: (articles: SavedArticleListEntryData[]) => void;
-}
-
-/**
- * Empty state component.
- */
-function SavedArticleListEmpty({ message }: { message: string }) {
-  return (
-    <div className="flex flex-col items-center justify-center py-12 text-center">
-      <svg
-        className="mb-4 h-12 w-12 text-zinc-400 dark:text-zinc-500"
-        fill="none"
-        viewBox="0 0 24 24"
-        stroke="currentColor"
-        aria-hidden="true"
-      >
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          strokeWidth={1.5}
-          d="M5 5a2 2 0 012-2h10a2 2 0 012 2v16l-7-3.5L5 21V5z"
-        />
-      </svg>
-      <p className="text-sm text-zinc-500 dark:text-zinc-400">{message}</p>
-    </div>
-  );
-}
-
-/**
- * Error state component.
- */
-function SavedArticleListError({ message, onRetry }: { message: string; onRetry: () => void }) {
-  return (
-    <div className="flex flex-col items-center justify-center py-12 text-center">
-      <svg
-        className="mb-4 h-12 w-12 text-red-400 dark:text-red-500"
-        fill="none"
-        viewBox="0 0 24 24"
-        stroke="currentColor"
-        aria-hidden="true"
-      >
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          strokeWidth={1.5}
-          d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z"
-        />
-      </svg>
-      <p className="mb-4 text-sm text-zinc-500 dark:text-zinc-400">{message}</p>
-      <button
-        onClick={onRetry}
-        className="rounded-md bg-zinc-900 px-4 py-2 text-sm font-medium text-white transition-colors hover:bg-zinc-800 dark:bg-zinc-50 dark:text-zinc-900 dark:hover:bg-zinc-200"
-      >
-        Try again
-      </button>
-    </div>
-  );
-}
-
-/**
- * Loading more indicator shown at bottom during pagination.
- */
-function LoadingMore() {
-  return (
-    <div
-      className="flex items-center justify-center py-4"
-      role="status"
-      aria-label="Loading more articles"
-    >
-      <svg
-        className="h-5 w-5 animate-spin text-zinc-500"
-        xmlns="http://www.w3.org/2000/svg"
-        fill="none"
-        viewBox="0 0 24 24"
-        aria-hidden="true"
-      >
-        <circle
-          className="opacity-25"
-          cx="12"
-          cy="12"
-          r="10"
-          stroke="currentColor"
-          strokeWidth="4"
-        />
-        <path
-          className="opacity-75"
-          fill="currentColor"
-          d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
-        />
-      </svg>
-      <span className="sr-only">Loading more articles...</span>
-    </div>
-  );
 }
 
 /**
@@ -253,7 +168,7 @@ export function SavedArticleList({
   // Error state
   if (isError) {
     return (
-      <SavedArticleListError
+      <ArticleListError
         message={error?.message ?? "Failed to load articles"}
         onRetry={() => refetch()}
       />
@@ -262,7 +177,7 @@ export function SavedArticleList({
 
   // Empty state
   if (allArticles.length === 0) {
-    return <SavedArticleListEmpty message={emptyMessage} />;
+    return <ArticleListEmpty message={emptyMessage} icon={<BookmarkEmptyIcon />} />;
   }
 
   return (
@@ -280,14 +195,10 @@ export function SavedArticleList({
       <div ref={loadMoreRef} className="h-1" />
 
       {/* Loading indicator */}
-      {isFetchingNextPage && <LoadingMore />}
+      {isFetchingNextPage && <ArticleListLoadingMore label="Loading more articles..." />}
 
       {/* End of list indicator */}
-      {!hasNextPage && allArticles.length > 0 && (
-        <p className="py-4 text-center text-sm text-zinc-400 dark:text-zinc-500">
-          No more articles
-        </p>
-      )}
+      {!hasNextPage && allArticles.length > 0 && <ArticleListEnd message="No more articles" />}
     </div>
   );
 }
